rename ui folder to components
