+++ src/app/components/modal/modal.css
