+++ src/app/request/update-winners-modal.js
@@ -4,7 +4,7 @@ import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import * as actions from '../../ducks/bounties';
 import { BOUNTY_ABI } from '../../utils/constants';
-import Modal from '../ui/modal';
+import Modal from '../components/modal';
 
 const bemify = block => (elem = '', modifier = '') => (
   `${block}${elem && '__' + elem}${modifier && '--' + modifier}`
@@ -83,7 +83,7 @@ class UpdateWinnersModal extends Component {
 
         this.props.onTxSubmit(data);
         this.props.onClose();
-      })   
+      })
     }
   }
